chore: fix release workflow indentation and update coordinates

+++ android-app/app/src/main/java/com/inputblocker/app/Region.kt
@@ -3,8 +3,8 @@ package com.inputblocker.app
 import java.io.Serializable
 
 data class Region(
-    var x1: Int = 0,
-    var y1: Int = 0,
-    var x2: Int = 0,
-    var y2: Int = 0
+    var x1: Float = 0f,
+    var y1: Float = 0f,
+    var x2: Float = 0f,
+    var y2: Float = 0f
 ) : Serializable
